server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
boto3==1.35.14
openai==1.45.0
